tests/example_test.py
import pytest
import allure

@allure.feature('Authentication')
@allure.story('User Login')
@allure.severity(allure.severity_level.CRITICAL)
def test_login_functionality():
    """Test user login with valid credentials"""
    with allure.step('Given a user with valid credentials'):
        username = "test_user"
        password = "test_pass"
    
    with allure.step('When the user attempts to login'):
        # Your login logic here
        login_successful = True  # Replace with actual login attempt
    
    with allure.step('Then the login should be successful'):
        assert login_successful, "Login failed"

@allure.feature('User Management')
@allure.story('User Profile')
def test_profile_update():
    with allure.step('Given a logged-in user'):
        # Setup code
        pass
    
    with allure.step('When updating profile information'):
        with allure.attachment_type.TEXT.attach('Profile Data', 'name: John Doe'):
            # Update profile logic
            update_successful = True
    
    with allure.step('Then the profile should be updated'):
        assert update_successful, "Profile update failed" 