requirements.txt
robotframework==6.1.1
robotframework-databaselibrary==1.2.4
robotframework-requests==0.9.6
pyodbc==4.0.39
allure-robotframework==2.13.2 
allure-pytest>=2.13.2
allure-python-commons>=2.13.2
pytest>=7.4.0 